+++ components/amorphic/lib/session/compressSessionData.js
@@ -1,21 +1,36 @@
 'use strict';
 
 // TODO: Make a SessionUtils
-
 let AmorphicContext = require('../AmorphicContext');
 let zlib = require('zlib');
+let statsdUtils = require('supertype').StatsdHelper;
 
-/**
- * Purpose unknown
- *
- * @param {unknown} data unknown
- *
- * @returns {unknown} unknown
+/*
+ * Compress session data
  */
 function compressSessionData(data) {
+    let compressSessionDataStartTime = process.hrtime();
+
     let amorphicOptions = AmorphicContext.amorphicOptions;
+
     if (amorphicOptions.compressSession) {
-        return zlib.deflateSync(data);
+        try {
+            let sessionData = zlib.deflateSync(data);
+
+            statsdUtils.computeTimingAndSend(
+                compressSessionDataStartTime,
+                'amorphic.session.compress_session_data.response_time',
+                { result: 'success' });
+
+            return sessionData;
+        } catch (e) {
+            statsdUtils.computeTimingAndSend(
+                compressSessionDataStartTime,
+                'amorphic.session.compress_session_data.response_time',
+                { result: 'failure' });
+
+            throw e;
+        }
     }
 
     return data;